JsonRpcClient: Check in Actindo's client
This is a working RPC client that Actindo gave to us during a workshop.
It implements a slightly non-standard version of JSON RPC 2.0 that uses
an `auth` field on the JSON RPC object in order to pass along an
authentication token.

Actindo doesn't maintain the client, so since we're taking ownership,
I've reformatted (whitespace, comments, and string literals only) to
match our standard style. My changes thus far in no way alter the
behavior. You can ignore this commit.