Schema: Add SchemaBuilder and login schema
The best option I've been able to find for documenting JSON RPC
interfaces is to define the schema of the valid requests and responses.
Using the JSON Schema specification and swaggest/php-code-builder, a
library for validating and hydrating JSON objects against a schema, we
can document the Actindo JSON RPC methods, their arguments, and their
expected responses.

Because the JSON RPC "envelope" is always the same except for the
request method, we opt for a shortcut in defining the requests. We just
document the single object argument that the Actindo API takes, rather
than the full request, which would always have top-level properties like
`method`, `args`, and `jsonrpc`.

Because JSON Schema isn't tied to a particular protocol (like JSON RPC),
there isn't a standard way to specify a request method and the expected
response. We use two custom schema fields to create our own minimal
"standard" for specifying requests and responses:

- `_method` specifies the JSON RPC method that expects the schema object
  as an argument.
- `_response_class` specifies the fully-qualified class name of a class
  that can validate and hydrate the response received from the API.

We have a `SchemaBuilder` class that uses swaggest/php-code-builder to
generate PHP classes corresponding to schemas under the `schema`
directory. The builder looks for these two custom properties when
building the PHP class corresponding to a request JSON object (a schema
whose path ends in `*.request.json`).

Our client will use the generated PHP classes and the custom properties
listed above to validate the requests it sends and the responses it
receives.