Client: Add Client, Request, Response
The `Client` class is the primary interface to interacting with the
Actindo PIM API. Internally, it builds requests validated by JSON Schema
via swagges/php-code-builder, uses the `JsonRpcClient` client to send
the validated argument to the correct method, and hydrates the
responses, again validating with JSON Schema.

Requests and Responses are abstracted from their bodies so that we can
instantiate different kinds of requests and responses (e.g., a request
built around JSON Schema, a more custom request, a response that
provides a pagination interface, and so on).