Client: Add method to list attribute sets
And to look up the identifier for the base attribute set, which is
required in order to modify product attributes. Note that there are many
attribute sets (they're like the properties of different classes), but
there should be only one PIM base set.

The API has a number of "list" requests, many (all?) of which take the
same pagination and filtering arguments. Add an interface to facilitate
specifying pagination and filters.

We're generally only adding what we need as we go, so there will be more
filter operations and accepted types for the filter value in the future,
and we'll probably expand the pagination interface to make it easy to
iterate through result sets.

Update the examples to demonstrate fetching the base attribute set id.