SchemaBuilder: Inject our own ClassStructure
It turns out that the `ClassStructure` that Swaggest provides does not
and cannot be made to complain if you try to get/set a property that
isn't defined by the backing schema. We would really like that
complaining to happen. To work around this, we inject our own base class
for the classes that get generated to represent objects conforming to
a schema.

Although we want to catch our own typos, we don't want to lock down the
responses that we receive from the API and throw an exception if they
add a new property. To handle that, we override `import()` to
temporarily turn off property validation on set.

In addition to adding validation and import logic, we also augment our
own `ClassStructure` with some convenient methods for validating and
exporting its data to PHP primitives and JSON.

Finally, this adds/fixes up some comments around class generation,
because there's quite a bit going on there.