Composer: Use our own php-code-builder
There's a bug on the master php-code-builder that throws a notice/error
when trying to generate a class constant for a JSON schema enum value
that doesn't have any valid constant characters. Depending on one's PHP
settings, this causes an error when trying to generate schema classes.
Use our own forked version of the branch that fixes the bug until it's
merged into master.

The specific issue that exposed the bug was the `=` value for the
`operation` property of the `filter` schema.