SchemaBuilder: Create output directory on demand
The schema code generator always completely wipes the directory that it
writes to before generating schema classes, so it's a pain to keep the
empty directory there with a `.gitkeep`, `.gitignore`, and so on.
Generating schema classes deletes the hidden file, and then you have to
be careful not to commit the deletion.

Instead, before beginning code generation, create the directory if it
doesn't exist.