Demo: Update example environment file
Depending on your file system, `ENV` can resolve to `/usr/bin/env`,
which prints your environment, so even if there's a file named `ENV` in
the current directory, `. ENV` doesn't source that file. `. ./ENV`
works, but it's easy to forget the `./`, and the error you get isn't
straightforward.

So let's just use a different name that doesn't conflict with (as far as
I know) any utilities and update the instructions accordingly.